split camera2basicfragment to three classes
