+++ app/src/main/java/com/digitalvotingpass/camera/Camera2BasicFragment.java
@@ -12,34 +12,21 @@ package com.digitalvotingpass.camera;/*
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
+ * This file was modified by the Digital Voting Pass group. (https://github.com/digital-voting-pass)
  */
 
 import android.Manifest;
 import android.app.Activity;
 import android.app.Fragment;
-import android.content.Context;
 import android.content.Intent;
 import android.content.pm.PackageManager;
-import android.content.res.Configuration;
 import android.graphics.Bitmap;
-import android.graphics.ImageFormat;
 import android.graphics.Matrix;
-import android.graphics.Point;
-import android.graphics.Rect;
 import android.graphics.RectF;
 import android.graphics.SurfaceTexture;
 import android.graphics.Typeface;
-import android.hardware.camera2.CameraAccessException;
-import android.hardware.camera2.CameraCaptureSession;
-import android.hardware.camera2.CameraCharacteristics;
-import android.hardware.camera2.CameraDevice;
-import android.hardware.camera2.CameraManager;
-import android.hardware.camera2.CaptureRequest;
-import android.hardware.camera2.params.MeteringRectangle;
-import android.hardware.camera2.params.StreamConfigurationMap;
 import android.os.Bundle;
-import android.os.Handler;
-import android.os.HandlerThread;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v13.app.FragmentCompat;
@@ -68,17 +55,29 @@ import com.digitalvotingpass.ocrscanner.TesseractOCR;
 import com.digitalvotingpass.utilities.ErrorDialog;
 
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.Comparator;
 import java.util.List;
 import java.util.concurrent.Semaphore;
-import java.util.concurrent.TimeUnit;
 
-public class Camera2BasicFragment extends Fragment
-        implements FragmentCompat.OnRequestPermissionsResultCallback {
+public class Camera2BasicFragment extends Fragment implements FragmentCompat.OnRequestPermissionsResultCallback {
+    // tag for the log and the error dialog
+    private static final String TAG = "Camera2BasicFragment";
+    private static final String FRAGMENT_DIALOG = "dialog";
 
     private static final int DELAY_BETWEEN_OCR_THREADS_MILLIS = 500;
+    private List<TesseractOCR> tesseractThreads = new ArrayList<>();
+    private boolean resultFound = false;
+    private Runnable scanningTakingLongTimeout = new Runnable() {
+        @Override
+        public void run() {
+            getActivity().runOnUiThread(new Runnable() {
+                @Override
+                public void run() {
+                    manualInput.setVisibility(View.VISIBLE);
+                    overlay.setMargins(0,0,0,infoText.getHeight() + manualInput.getHeight());
+                }
+            });
+        }
+    };
 
     private ImageView scanSegment;
     private Overlay overlay;
@@ -86,23 +85,10 @@ public class Camera2BasicFragment extends Fragment
     private TextView infoText;
     private View controlPanel;
 
-    private List<TesseractOCR> tesseractThreads = new ArrayList<>();
-    /**
-     * If this is defined and > 0, use this amount of threads instead of dynamically determined amount.
-     */
-    private int ocrThreadNumberOverride;
-
-    /**
-     * Conversion from screen rotation to JPEG orientation.
-     */
+    // Conversion from screen rotation to JPEG orientation.
     private static final SparseIntArray ORIENTATIONS = new SparseIntArray();
     private static final int REQUEST_CAMERA_PERMISSION = 1;
     private static final int REQUEST_WRITE_PERMISSIONS = 3;
-
-    private static final String FRAGMENT_DIALOG = "dialog";
-
-    boolean mIsStateAlreadySaved = false;
-    boolean mPendingShowDialog = false;
 
     static {
         ORIENTATIONS.append(Surface.ROTATION_0, 90);
@@ -111,23 +97,11 @@ public class Camera2BasicFragment extends Fragment
         ORIENTATIONS.append(Surface.ROTATION_270, 180);
     }
 
+    boolean mIsStateAlreadySaved = false;
+    boolean mPendingShowDialog = false;
+
     // listener for detecting orientation changes
     private OrientationEventListener orientationListener = null;
-
-    /**
-     * Tag for the {@link Log}.
-     */
-    private static final String TAG = "Camera2BasicFragment";
-
-    /**
-     * Max preview width that is guaranteed by Camera2 API
-     */
-    private static final int MAX_PREVIEW_WIDTH = 1920;
-
-    /**
-     * Max preview height that is guaranteed by Camera2 API
-     */
-    private static final int MAX_PREVIEW_HEIGHT = 1080;
 
     /**
      * {@link TextureView.SurfaceTextureListener} handles several lifecycle events on a
@@ -139,7 +113,7 @@ public class Camera2BasicFragment extends Fragment
         @Override
         public void onSurfaceTextureAvailable(SurfaceTexture texture, int width, int height) {
             orientationListener.enable();
-            openCamera(width, height);
+            mCameraHandler.openCamera(width, height);
         }
 
         @Override
@@ -159,9 +133,9 @@ public class Camera2BasicFragment extends Fragment
     };
 
     /**
-     * ID of the current {@link CameraDevice}.
+     * A {@link Semaphore} to prevent the app from exiting before closing the camera.
      */
-    private String mCameraId;
+    public Semaphore mCameraOpenCloseLock = new Semaphore(1);
 
     /**
      * An {@link AutoFitTextureView} for camera preview.
@@ -169,197 +143,25 @@ public class Camera2BasicFragment extends Fragment
     private AutoFitTextureView mTextureView;
 
     /**
-     * A {@link CameraCaptureSession } for camera preview.
-     */
-    private CameraCaptureSession mCaptureSession;
-
-    /**
-     * A reference to the opened {@link CameraDevice}.
-     */
-    private CameraDevice mCameraDevice;
-
-    /**
      * The {@link android.util.Size} of camera preview.
      */
     private Size mPreviewSize;
-    private boolean resultFound = false;
-
-    private float secTillScanTimeout = 10;
-    private Runnable scanningTakingLongTimeout = new Runnable() {
-        @Override
-        public void run() {
-            getActivity().runOnUiThread(new Runnable() {
-                @Override
-                public void run() {
-                    manualInput.setVisibility(View.VISIBLE);
-                    overlay.setMargins(0,0,0,infoText.getHeight() + manualInput.getHeight());
-                }
-            });
-        }
-    };
 
     /**
-     * Method for delivering correct MRZ when found. This method returns the MRZ as result data and
-     * then exits the activity. This method is synchronized and checks for a boolean to make sure
-     * it is only executed once in this fragments lifetime.
-     * @param mrz Mrz
+     * Handler for the connection with the camera
      */
-    public synchronized void scanResultFound(final Mrz mrz) {
-        if (!resultFound) {
-            for (TesseractOCR thread : tesseractThreads) {
-                thread.stopping = true;
-            }
-            Intent returnIntent = new Intent();
-            DocumentData data = mrz.getPrettyData();
-            returnIntent.putExtra("result", data);
-            getActivity().setResult(Activity.RESULT_OK, returnIntent);
-            resultFound = true;
-            finishActivity();
-        }
-    }
+    private CameraHandler mCameraHandler;
 
-    private void finishActivity() {
-        getActivity().finish();
-    }
 
-    /**
-     * {@link CameraDevice.StateCallback} is called when {@link CameraDevice} changes its state.
-     */
-    private final CameraDevice.StateCallback mStateCallback = new CameraDevice.StateCallback() {
-
-        @Override
-        public void onOpened(@NonNull CameraDevice cameraDevice) {
-            // This method is called when the camera is opened.  We start camera preview here.
-            mCameraOpenCloseLock.release();
-            mCameraDevice = cameraDevice;
-            createCameraPreviewSession();
-        }
-
-        @Override
-        public void onDisconnected(@NonNull CameraDevice cameraDevice) {
-            mCameraOpenCloseLock.release();
-            cameraDevice.close();
-            mCameraDevice = null;
-        }
-
-        @Override
-        public void onError(@NonNull CameraDevice cameraDevice, int error) {
-            mCameraOpenCloseLock.release();
-            cameraDevice.close();
-            mCameraDevice = null;
-            Activity activity = getActivity();
-            if (null != activity) {
-                activity.finish();
-            }
-        }
-    };
-
-    /**
-     * An additional thread for running tasks that shouldn't block the UI.
-     */
-    private HandlerThread mBackgroundThread;
-
-    /**
-     * A {@link Handler} for running tasks in the background.
-     * Runs camera preview updater.
-     */
-    private Handler mBackgroundHandler;
-
-    /**
-     * {@link CaptureRequest.Builder} for the camera preview
-     */
-    private CaptureRequest.Builder mPreviewRequestBuilder;
-
-    /**
-     * {@link CaptureRequest} generated by {@link #mPreviewRequestBuilder}
-     */
-    private CaptureRequest mPreviewRequest;
-
-    /**
-     * A {@link Semaphore} to prevent the app from exiting before closing the camera.
-     */
-    private Semaphore mCameraOpenCloseLock = new Semaphore(1);
-
-    /**
-     * Whether the current camera device supports Flash or not.
-     */
-    private boolean mFlashSupported;
-
-    /**
-     * Orientation of the camera sensor
-     */
-    private int mSensorOrientation;
-
-    /**
-     * Shows a {@link Toast} on the UI thread.
-     *
-     * @param text The message to show
-     */
-    private void showToast(final String text) {
-        final Activity activity = getActivity();
-        if (activity != null) {
-            activity.runOnUiThread(new Runnable() {
-                @Override
-                public void run() {
-                    Toast.makeText(activity, text, Toast.LENGTH_SHORT).show();
-                }
-            });
-        }
-    }
-
-    /**
-     * Given {@code choices} of {@code Size}s supported by a camera, choose the smallest one that
-     * is at least as large as the respective texture view size, and that is at most as large as the
-     * respective max size, and whose aspect ratio matches with the specified value. If such size
-     * doesn't exist, choose the largest one that is at most as large as the respective max size,
-     * and whose aspect ratio matches with the specified value.
-     *
-     * @param choices           The list of sizes that the camera supports for the intended output
-     *                          class
-     * @param textureViewWidth  The width of the texture view relative to sensor coordinate
-     * @param textureViewHeight The height of the texture view relative to sensor coordinate
-     * @param maxWidth          The maximum width that can be chosen
-     * @param maxHeight         The maximum height that can be chosen
-     * @param aspectRatio       The aspect ratio
-     * @return The optimal {@code Size}, or an arbitrary one if none were big enough
-     */
-    private static Size chooseOptimalSize(Size[] choices, int textureViewWidth,
-                                          int textureViewHeight, int maxWidth, int maxHeight, Size aspectRatio) {
-
-        // Collect the supported resolutions that are at least as big as the preview Surface
-        List<Size> bigEnough = new ArrayList<>();
-        // Collect the supported resolutions that are smaller than the preview Surface
-        List<Size> notBigEnough = new ArrayList<>();
-        int w = aspectRatio.getWidth();
-        int h = aspectRatio.getHeight();
-        for (Size option : choices) {
-            if (option.getWidth() <= maxWidth && option.getHeight() <= maxHeight &&
-                    option.getHeight() == option.getWidth() * h / w) {
-                if (option.getWidth() >= textureViewWidth &&
-                        option.getHeight() >= textureViewHeight) {
-                    bigEnough.add(option);
-                } else {
-                    notBigEnough.add(option);
-                }
-            }
-        }
-
-        // Pick the smallest of those big enough. If there is no one big enough, pick the
-        // largest of those not big enough.
-        if (bigEnough.size() > 0) {
-            return Collections.min(bigEnough, new CompareSizesByArea());
-        } else if (notBigEnough.size() > 0) {
-            return Collections.max(notBigEnough, new CompareSizesByArea());
-        } else {
-            Log.e(TAG, "Couldn't find any suitable preview size");
-            return choices[0];
-        }
-    }
 
     public static Camera2BasicFragment newInstance() {
         return new Camera2BasicFragment();
     }
 
+    /**
+     * Handles the setup that can start when the fragment is created.
+     * @param savedInstanceState
+     */
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -369,12 +171,14 @@ public class Camera2BasicFragment extends Fragment
             }
         };
         int threadsToStart = Runtime.getRuntime().availableProcessors() / 2;
-        if (ocrThreadNumberOverride > 0) {
-            threadsToStart = ocrThreadNumberOverride;
-        }
         createOCRThreads(threadsToStart);
+        mCameraHandler = new CameraHandler(this);
     }
 
+    /**
+     * Create the threads where the OCR will run on.
+     * @param amount
+     */
     private void createOCRThreads(int amount) {
         for (int i = 0; i < amount; i++) {
             tesseractThreads.add(new TesseractOCR("Thread no " + i, this, getActivity().getAssets()));
@@ -388,6 +192,9 @@ public class Camera2BasicFragment extends Fragment
         return inflater.inflate(R.layout.fragment_camera2_basic, container, false);
     }
 
+    /**
+     * Setup the layout and setup the actions associated with the button.
+     */
     @Override
     public void onViewCreated(final View view, final Bundle savedInstanceState) {
         mTextureView = (AutoFitTextureView) view.findViewById(R.id.texture);
@@ -425,13 +232,13 @@ public class Camera2BasicFragment extends Fragment
     @Override
     public void onResume() {
         super.onResume();
-        startBackgroundThread();
+        mCameraHandler.startBackgroundThread();
         // When the screen is turned off and turned back on, the SurfaceTexture is already
         // available, and "onSurfaceTextureAvailable" will not be called. In that case, we can open
         // a camera and start preview from here (otherwise, we wait until the surface is ready in
         // the SurfaceTextureListener).
         if (mTextureView.isAvailable()) {
-            openCamera(mTextureView.getWidth(), mTextureView.getHeight());
+            mCameraHandler.openCamera(mTextureView.getWidth(), mTextureView.getHeight());
         } else {
             mTextureView.setSurfaceTextureListener(mSurfaceTextureListener);
         }
@@ -446,19 +253,23 @@ public class Camera2BasicFragment extends Fragment
                 showInfoDialog(R.string.storage_permission_explanation);
             }
         } else {
-            startTesseractThreads();
+            startOCRThreads();
         }
     }
 
-    private void showInfoDialog(int stringid) {
-        ErrorDialog.newInstance(getString(stringid))
+    /**
+     * Displays an information dialog with the given string.
+     * @param stringId
+     */
+    public void showInfoDialog(int stringId) {
+        ErrorDialog.newInstance(getString(stringId))
                 .show(getChildFragmentManager(), FRAGMENT_DIALOG);
     }
 
     @Override
     public void onPause() {
-        closeCamera();
-        stopBackgroundThread();
+        mCameraHandler.closeCamera();
+        mCameraHandler.stopBackgroundThread();
         stopTesseractThreads();
         mIsStateAlreadySaved = true;
         super.onPause();
@@ -474,10 +285,9 @@ public class Camera2BasicFragment extends Fragment
         super.onStop();
     }
 
-    private void requestCameraPermission() {
+    public void requestCameraPermission() {
         if (FragmentCompat.shouldShowRequestPermissionRationale(this, Manifest.permission.CAMERA)) {
-            ErrorDialog.newInstance(getString(R.string.ocr_camera_permission_explanation))
-                    .show(getChildFragmentManager(), FRAGMENT_DIALOG);
+            showInfoDialog(R.string.ocr_camera_permission_explanation);
         } else {
             FragmentCompat.requestPermissions(this, new String[]{Manifest.permission.CAMERA},
                     REQUEST_CAMERA_PERMISSION);
@@ -486,8 +296,7 @@ public class Camera2BasicFragment extends Fragment
 
     private void requestStoragePermissions() {
         if (FragmentCompat.shouldShowRequestPermissionRationale(this, Manifest.permission.WRITE_EXTERNAL_STORAGE)) {
-            ErrorDialog.newInstance(getString(R.string.storage_permission_explanation))
-                    .show(getChildFragmentManager(), FRAGMENT_DIALOG);
+            showInfoDialog(R.string.storage_permission_explanation);
         } else {
             FragmentCompat.requestPermissions(this, new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE}, REQUEST_WRITE_PERMISSIONS);
         }
@@ -501,8 +310,7 @@ public class Camera2BasicFragment extends Fragment
                 if(mIsStateAlreadySaved){
                     mPendingShowDialog = true;
                 } else {
-                    ErrorDialog.newInstance(getString(R.string.ocr_camera_permission_explanation))
-                            .show(getChildFragmentManager(), FRAGMENT_DIALOG);
+                    showInfoDialog(R.string.ocr_camera_permission_explanation);
                 }
             }
         } else if (requestCode == REQUEST_WRITE_PERMISSIONS) {
@@ -510,8 +318,7 @@ public class Camera2BasicFragment extends Fragment
                 if(mIsStateAlreadySaved){
                     mPendingShowDialog = true;
                 } else {
-                    ErrorDialog.newInstance(getString(R.string.storage_permission_explanation))
-                            .show(getChildFragmentManager(), FRAGMENT_DIALOG);
+                    showInfoDialog(R.string.storage_permission_explanation);
                 }
             }
         } else {
@@ -520,175 +327,9 @@ public class Camera2BasicFragment extends Fragment
     }
 
     /**
-     * Sets up member variables related to camera.
-     *
-     * @param width  The width of available size for camera preview
-     * @param height The height of available size for camera preview
+     * Start the threads that will run the OCR scanner.
      */
-    private void setUpCameraOutputs(int width, int height) {
-        Activity activity = getActivity();
-        CameraManager manager = (CameraManager) activity.getSystemService(Context.CAMERA_SERVICE);
-        try {
-            for (String cameraId : manager.getCameraIdList()) {
-                CameraCharacteristics characteristics
-                        = manager.getCameraCharacteristics(cameraId);
-
-                // We don't use a front facing camera in this sample.
-                Integer facing = characteristics.get(CameraCharacteristics.LENS_FACING);
-                if (facing != null && facing == CameraCharacteristics.LENS_FACING_FRONT) {
-                    continue;
-                }
-
-                StreamConfigurationMap map = characteristics.get(
-                        CameraCharacteristics.SCALER_STREAM_CONFIGURATION_MAP);
-                if (map == null) {
-                    continue;
-                }
-
-                // For still image captures, we use the largest available size.
-                Size largest = Collections.max(
-                        Arrays.asList(map.getOutputSizes(ImageFormat.JPEG)),
-                        new CompareSizesByArea());
-
-                // Find out if we need to swap dimension to get the preview size relative to sensor
-                // coordinate.
-                int displayRotation = activity.getWindowManager().getDefaultDisplay().getRotation();
-                //noinspection ConstantConditions
-                mSensorOrientation = characteristics.get(CameraCharacteristics.SENSOR_ORIENTATION);
-                boolean swappedDimensions = false;
-                switch (displayRotation) {
-                    case Surface.ROTATION_0:
-                    case Surface.ROTATION_90:
-                        if (mSensorOrientation == 90 || mSensorOrientation == 270) {
-                            swappedDimensions = true;
-                        }
-                        break;
-                    case Surface.ROTATION_180:
-                    case Surface.ROTATION_270:
-                        if (mSensorOrientation == 0 || mSensorOrientation == 180) {
-                            swappedDimensions = true;
-                        }
-                        break;
-                    default:
-                        Log.e(TAG, "Display rotation is invalid: " + displayRotation);
-                }
-
-                Point displaySize = new Point();
-                activity.getWindowManager().getDefaultDisplay().getSize(displaySize);
-                int rotatedPreviewWidth = width;
-                int rotatedPreviewHeight = height;
-                int maxPreviewWidth = displaySize.x;
-                int maxPreviewHeight = displaySize.y;
-
-                if (swappedDimensions) {
-                    rotatedPreviewWidth = height;
-                    rotatedPreviewHeight = width;
-                    maxPreviewWidth = displaySize.y;
-                    maxPreviewHeight = displaySize.x;
-                }
-
-                if (maxPreviewWidth > MAX_PREVIEW_WIDTH) {
-                    maxPreviewWidth = MAX_PREVIEW_WIDTH;
-                }
-
-                if (maxPreviewHeight > MAX_PREVIEW_HEIGHT) {
-                    maxPreviewHeight = MAX_PREVIEW_HEIGHT;
-                }
-                // Danger, W.R.! Attempting to use too large a preview size could  exceed the camera
-                // bus' bandwidth limitation, resulting in gorgeous previews but the storage of
-                // garbage capture data.
-                mPreviewSize = chooseOptimalSize(map.getOutputSizes(SurfaceTexture.class),
-                        rotatedPreviewWidth, rotatedPreviewHeight, maxPreviewWidth,
-                        maxPreviewHeight, largest);
-
-                int orientation = getResources().getConfiguration().orientation;
-                if (orientation == Configuration.ORIENTATION_LANDSCAPE) {
-                    mTextureView.setAspectRatio(
-                            mPreviewSize.getWidth(), mPreviewSize.getHeight());
-                } else {
-                    mTextureView.setAspectRatio(
-                            mPreviewSize.getHeight(), mPreviewSize.getWidth());
-                }
-
-                // Check if the flash is supported.
-                Boolean available = characteristics.get(CameraCharacteristics.FLASH_INFO_AVAILABLE);
-                mFlashSupported = available == null ? false : available;
-
-                mCameraId = cameraId;
-                return;
-            }
-        } catch (CameraAccessException e) {
-            e.printStackTrace();
-        } catch (NullPointerException e) {
-            // Currently an NPE is thrown when the Camera2API is used but not supported on the
-            // device this code runs.
-            ErrorDialog.newInstance(getString(R.string.ocr_camera_error))
-                    .show(getChildFragmentManager(), FRAGMENT_DIALOG);
-        }
-    }
-
-    /**
-     * Opens the camera specified by {@link Camera2BasicFragment#mCameraId}.
-     */
-    private void openCamera(int width, int height) {
-        if (ContextCompat.checkSelfPermission(getActivity(), Manifest.permission.CAMERA)
-                != PackageManager.PERMISSION_GRANTED) {
-            requestCameraPermission();
-            return;
-        }
-        if (ContextCompat.checkSelfPermission(getActivity(), Manifest.permission.WRITE_EXTERNAL_STORAGE)
-                != PackageManager.PERMISSION_GRANTED) {
-            Log.e(TAG, "Cant start Camera due to no data permissions.");
-            return;
-        }
-        setUpCameraOutputs(width, height);
-        configureTransform(width, height);
-        Activity activity = getActivity();
-        CameraManager manager = (CameraManager) activity.getSystemService(Context.CAMERA_SERVICE);
-        try {
-            if (!mCameraOpenCloseLock.tryAcquire(2500, TimeUnit.MILLISECONDS)) {
-                throw new RuntimeException("Time out waiting to lock camera opening.");
-            }
-            manager.openCamera(mCameraId, mStateCallback, mBackgroundHandler);
-        } catch (CameraAccessException e) {
-            e.printStackTrace();
-        } catch (InterruptedException e) {
-            throw new RuntimeException("Interrupted while trying to lock camera opening.", e);
-        }
-    }
-
-    /**
-     * Closes the current {@link CameraDevice}.
-     */
-    private void closeCamera() {
-        try {
-            mCameraOpenCloseLock.acquire();
-            if (null != mCaptureSession) {
-                mCaptureSession.close();
-                mCaptureSession = null;
-            }
-            if (null != mCameraDevice) {
-                mCameraDevice.close();
-                mCameraDevice = null;
-            }
-        } catch (InterruptedException e) {
-            throw new RuntimeException("Interrupted while trying to lock camera closing.", e);
-        } finally {
-            mCameraOpenCloseLock.release();
-        }
-    }
-
-    /**
-     * Starts a background thread and its {@link Handler}.
-     */
-    private void startBackgroundThread() {
-        mBackgroundThread = new HandlerThread("CameraBackground");
-        mBackgroundThread.start();
-        mBackgroundHandler = new Handler(mBackgroundThread.getLooper());
-        mBackgroundHandler.postDelayed(scanningTakingLongTimeout, (long) (secTillScanTimeout * 1000));
-    }
-
-    private void startTesseractThreads() {
+    private void startOCRThreads() {
         if (ContextCompat.checkSelfPermission(getActivity(), Manifest.permission.WRITE_EXTERNAL_STORAGE)
                 != PackageManager.PERMISSION_GRANTED) {
             requestStoragePermissions();
@@ -707,20 +348,7 @@ public class Camera2BasicFragment extends Fragment
         }
     }
 
-    /**
-     * Stops the background thread and its {@link Handler}.
-     */
-    private void stopBackgroundThread() {
-        mBackgroundHandler.removeCallbacks(scanningTakingLongTimeout);
-        mBackgroundThread.quitSafely();
-        try {
-            mBackgroundThread.join();
-            mBackgroundThread = null;
-            mBackgroundHandler = null;
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-    }
+
 
     private void stopTesseractThreads() {
         for (TesseractOCR ocr : tesseractThreads) {
@@ -728,79 +356,26 @@ public class Camera2BasicFragment extends Fragment
         }
     }
 
+
+
     /**
-     * Creates a new {@link CameraCaptureSession} for camera preview.
+     * Method for delivering correct MRZ when found. This method returns the MRZ as result data and
+     * then exits the activity. This method is synchronized and checks for a boolean to make sure
+     * it is only executed once in this fragments lifetime.
+     * @param mrz Mrz
      */
-    private void createCameraPreviewSession() {
-        try {
-            SurfaceTexture texture = mTextureView.getSurfaceTexture();
-            assert texture != null;
-
-            // We configure the size of default buffer to be the size of camera preview we want.
-            texture.setDefaultBufferSize(mPreviewSize.getWidth(), mPreviewSize.getHeight());
-
-            // This is the output Surface we need to start preview.
-            Surface surface = new Surface(texture);
-
-            // We set up a CaptureRequest.Builder with the output Surface.
-            mPreviewRequestBuilder
-                    = mCameraDevice.createCaptureRequest(CameraDevice.TEMPLATE_PREVIEW);
-            mPreviewRequestBuilder.addTarget(surface);
-            // Here, we create a CameraCaptureSession for camera preview.
-            mCameraDevice.createCaptureSession(Arrays.asList(surface),
-                    new CameraCaptureSession.StateCallback() {
-
-                        @Override
-                        public void onConfigured(@NonNull CameraCaptureSession cameraCaptureSession) {
-                            // The camera is already closed
-                            if (null == mCameraDevice) {
-                                return;
-                            }
-                            // When the session is ready, we start displaying the preview.
-                            mCaptureSession = cameraCaptureSession;
-                            try {
-                                MeteringRectangle meteringRectangle=new MeteringRectangle(getScanRect(),
-                                        MeteringRectangle.METERING_WEIGHT_MAX);
-                                MeteringRectangle[] meteringRectangleArr={meteringRectangle};
-
-                                // Auto focus should be continuous for camera preview.
-                                mPreviewRequestBuilder.set(CaptureRequest.CONTROL_AF_REGIONS,
-                                        meteringRectangleArr);
-                                // Flash is automatically enabled when necessary.
-                                setAutoFlash(mPreviewRequestBuilder);
-
-                                // Finally, we start displaying the camera preview.
-                                mPreviewRequest = mPreviewRequestBuilder.build();
-                                if (!mIsStateAlreadySaved)
-                                    mCaptureSession.setRepeatingRequest(mPreviewRequest,
-                                            null, mBackgroundHandler);
-                            } catch (CameraAccessException e) {
-                                e.printStackTrace();
-                            }
-                        }
-
-                        @Override
-                        public void onConfigureFailed(
-                                @NonNull CameraCaptureSession cameraCaptureSession) {
-                            showToast("Failed");
-                        }
-                    }, null
-            );
-        } catch (CameraAccessException e) {
-            e.printStackTrace();
+    public synchronized void scanResultFound(final Mrz mrz) {
+        if (!resultFound) {
+            for (TesseractOCR thread : tesseractThreads) {
+                thread.stopping = true;
+            }
+            Intent returnIntent = new Intent();
+            DocumentData data = mrz.getPrettyData();
+            returnIntent.putExtra("result", data);
+            getActivity().setResult(Activity.RESULT_OK, returnIntent);
+            resultFound = true;
+            getActivity().finish();
         }
-    }
-
-    /**
-     * Get the scan rectangle.
-     * @return The rectangle.
-     */
-    private Rect getScanRect() {
-        int startX = (int) scanSegment.getX();
-        int startY = (int) scanSegment.getY();
-        int width = scanSegment.getWidth();
-        int length = scanSegment.getHeight();
-        return new Rect(startX, startY, startX + width, startY+length);
     }
 
     /**
@@ -811,7 +386,7 @@ public class Camera2BasicFragment extends Fragment
      * @param viewWidth  The width of `mTextureView`
      * @param viewHeight The height of `mTextureView`
      */
-    private void configureTransform(int viewWidth, int viewHeight) {
+    public void configureTransform(int viewWidth, int viewHeight) {
         Activity activity = getActivity();
         if (null == mTextureView || null == mPreviewSize || null == activity) {
             return;
@@ -834,95 +409,91 @@ public class Camera2BasicFragment extends Fragment
             matrix.postRotate(180, centerX, centerY);
         }
         mTextureView.setTransform(matrix);
-        overlay.setRect(getScanRect());
-    }
-
-    private void setAutoFlash(CaptureRequest.Builder requestBuilder) {
-        if (mFlashSupported) {
-            requestBuilder.set(CaptureRequest.CONTROL_AE_MODE,
-                    CaptureRequest.CONTROL_AE_MODE_ON_AUTO_FLASH);
-        }
+        overlay.setRect(CameraSupport.getScanRect(scanSegment));
     }
 
     /**
-     * Compares two {@code Size}s based on their areas.
+     * Extract a bitmap from the textureview of this fragment.
+     * @return
      */
-    static class CompareSizesByArea implements Comparator<Size> {
-
-        @Override
-        public int compare(Size lhs, Size rhs) {
-            // We cast here to ensure the multiplications won't overflow
-            return Long.signum((long) lhs.getWidth() * lhs.getHeight() -
-                    (long) rhs.getWidth() * rhs.getHeight());
-        }
-
-    }
-
     public Bitmap extractBitmap() {
         try {
             Bitmap bitmap = mTextureView.getBitmap();
-            int rotate = 0;
+            int rotate = Surface.ROTATION_0;
             switch (getActivity().getWindowManager().getDefaultDisplay().getRotation()) {
-                case 0:
+                case Surface.ROTATION_0:
                     rotate = 0;
                     break;
-                case 1:
+                case Surface.ROTATION_90:
                     rotate = 270;
                     break;
-                case 2:
+                case Surface.ROTATION_180:
                     rotate = 180;
                     break;
-                case 3:
+                case Surface.ROTATION_270:
                     rotate = 90;
                     break;
             }
-            if (rotate != 0) {
-                bitmap = rotateBitmap(bitmap, rotate);
+            if (rotate != Surface.ROTATION_0) {
+                bitmap = CameraSupport.rotateBitmap(bitmap, rotate);
             }
-            Bitmap croppedBitmap = cropBitmap(bitmap);
+            Bitmap croppedBitmap = CameraSupport.cropBitmap(bitmap, scanSegment);
 
-            final Bitmap b = getResizedBitmap(croppedBitmap, (int) (croppedBitmap.getWidth()), (int) (croppedBitmap.getHeight()));
-
-            return b;
+            return CameraSupport.getResizedBitmap(croppedBitmap, croppedBitmap.getWidth(), croppedBitmap.getHeight());
         } catch (Exception e) {
             e.printStackTrace();
             return null;
         }
+
     }
 
-    private Bitmap rotateBitmap(Bitmap bitmap, int degrees) {
-        int w = bitmap.getWidth();
-        int h = bitmap.getHeight();
-        // Setting pre rotate
-        Matrix mtx = new Matrix();
-        mtx.preRotate(degrees);
-        // Rotating Bitmap
-        Bitmap rotated = Bitmap.createBitmap(bitmap, 0, 0, w, h, mtx, true);
-        return Bitmap.createScaledBitmap(rotated, bitmap.getWidth(), bitmap.getHeight(), true);
+    public Size getPreviewSize() {
+        return mPreviewSize;
     }
 
-    public Bitmap getResizedBitmap(Bitmap bm, int newWidth, int newHeight) {
-        int width = bm.getWidth();
-        int height = bm.getHeight();
-        float scaleWidth = ((float) newWidth) / width;
-        float scaleHeight = ((float) newHeight) / height;
-        // CREATE A MATRIX FOR THE MANIPULATION
-        Matrix matrix = new Matrix();
-        // RESIZE THE BIT MAP
-        matrix.postScale(scaleWidth, scaleHeight);
-
-        // "RECREATE" THE NEW BITMAP
-        Bitmap resizedBitmap = Bitmap.createBitmap(
-                bm, 0, 0, width, height, matrix, false);
-        bm.recycle();
-        return resizedBitmap;
+    public AutoFitTextureView getTextureView() {
+        return mTextureView;
     }
 
-    private Bitmap cropBitmap(Bitmap bitmap) {
-        int startX = (int) scanSegment.getX();
-        int startY = (int) scanSegment.getY();
-        int width = (int) (scanSegment.getWidth());
-        int length = (int) (scanSegment.getHeight());
-        return Bitmap.createBitmap(bitmap, startX, startY, width, length);
+    public ImageView getScanSegment() {
+        return scanSegment;
+    }
+
+    public boolean isStateAlreadySaved() {
+        return mIsStateAlreadySaved;
+    }
+
+    public Runnable getScanningTakingLongTimeout() {
+        return scanningTakingLongTimeout;
+    }
+
+    public void setPreviewSize(Size size) {
+        mPreviewSize = size;
+    }
+
+    /**
+     * Sets the aspect ratio of the textureview
+     * @param width
+     * @param height
+     */
+    public void setAspectRatio(int width, int height) {
+        mTextureView.setAspectRatio(width, height);
+    }
+
+    /**
+     * Shows a {@link Toast} on the UI thread.
+     *
+     * @param text The message to show
+     */
+    public void showToast(final String text) {
+        final Activity activity = getActivity();
+        if (activity != null) {
+            activity.runOnUiThread(new Runnable() {
+                @Override
+                public void run() {
+                    Toast.makeText(activity, text, Toast.LENGTH_SHORT).show();
+                }
+            });
+        }
     }
 }
