+++ app/src/main/java/com/digitalvotingpass/digitalvotingpass/ResultActivity.java
@@ -14,7 +14,6 @@ import android.view.View;
 import android.view.ViewTreeObserver;
 import android.widget.Button;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.androidadvance.topsnackbar.TSnackbar;
 import com.digitalvotingpass.blockchain.BlockChain;
